Added FpsCounter, for measuring FPS
Added FpsCounter, for measuring FPS
    src/core/MainManager.cpp
    src/core/MainManager.h
    src/util/Averager.h
    src/util/FpsCounter.cpp (+)
    src/util/FpsCounter.h   (+)

+++ src/util/Averager.h
@@ -71,7 +71,7 @@ public:
   double getAverage() const
   {
     assert(N_ != 0);
-    return MX_n_ / N_;
+    return MX_n_ / static_cast<double>(N_);
   }
